% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiladOlad-package.R
\docType{package}
\name{MiladOlad-package}
\alias{MiladOlad}
\alias{MiladOlad-package}
\title{MiladOlad: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/MiladOlad}
  \item Report bugs at \url{https://github.com/f-dallolio/MiladOlad/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
